refactored project structure and added local invoker/debugger cli
