+++ cli/runner.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
+	"strings"
 	"syscall"
 )
 
-// RunMode represents how the Lambda function should be executed
 type RunMode int
 
 const (
@@ -17,20 +17,17 @@ const (
 	ModeGoRun
 )
 
-// Runner handles the execution of Lambda functions
 type Runner struct {
 	Mode       RunMode
 	Debug      bool
 	ServerPort string
 }
 
-// CommandConfig contains the parsed command configuration
 type CommandConfig struct {
 	LambdaPath string
 	LambdaArgs []string
 }
 
-// NewRunner creates a new Runner with the specified mode
 func NewRunner(mode RunMode, debug bool, serverPort string) *Runner {
 	return &Runner{
 		Mode:       mode,
@@ -43,28 +40,31 @@ func NewRunner(mode RunMode, debug bool, serverPort string) *Runner {
 // args should be the result of flag.Args() after flag.Parse()
 func (r *Runner) ParseArgs(args []string) (*CommandConfig, error) {
 	config := &CommandConfig{}
+
 	r.Debugf("Parsing arguments: %v", args)
+
+	if len(args) < 1 {
+		return nil, fmt.Errorf("missing lambda path argument")
+	}
+
 	switch r.Mode {
 	case ModeBinary:
-		if len(args) < 1 {
-			return nil, fmt.Errorf("missing lambda path argument")
-		}
+
 		config.LambdaPath = args[0]
 		if len(args) > 1 {
 			config.LambdaArgs = args[1:]
 		}
 
 	case ModeGoRun:
-		// For go run mode, we expect: -- lambda-path [args...]
-		if len(args) < 2 {
-			return nil, fmt.Errorf("missing lambda path argument (expected after '--')")
+		isGoFile := strings.HasSuffix(args[0], ".go")
+
+		if !isGoFile {
+			return nil, fmt.Errorf("invalid go file: %s", args[0])
 		}
-		// if args[0] != "--" {
-		// 	return nil, fmt.Errorf("expected '--' separator before lambda path, got: %s", args[0])
-		// }
+
 		config.LambdaPath = args[0]
-		if len(args) > 2 {
-			config.LambdaArgs = args[2:]
+		if len(args) > 1 {
+			config.LambdaArgs = args[1:]
 		}
 
 	default:
@@ -74,52 +74,35 @@ func (r *Runner) ParseArgs(args []string) (*CommandConfig, error) {
 	return config, nil
 }
 
-// CreateCommand creates an exec.Command based on the run mode
 func (r *Runner) CreateCommand(config *CommandConfig) (*exec.Cmd, error) {
 	var cmd *exec.Cmd
 
+	if _, err := os.Stat(config.LambdaPath); os.IsNotExist(err) {
+		return nil, fmt.Errorf("not found at %s", config.LambdaPath)
+	}
+
 	switch r.Mode {
 	case ModeBinary:
-		// Check if lambda binary exists
-		if _, err := os.Stat(config.LambdaPath); os.IsNotExist(err) {
-			return nil, fmt.Errorf("lambda binary not found at %s", config.LambdaPath)
-		}
 		cmd = exec.Command(config.LambdaPath, config.LambdaArgs...)
 
 	case ModeGoRun:
-		// Check if lambda source file exists
-		if _, err := os.Stat(config.LambdaPath); os.IsNotExist(err) {
-			return nil, fmt.Errorf("lambda source file not found at %s", config.LambdaPath)
-		}
 		// Construct args: go run <lambda-path> [lambda-args...]
 		args := append([]string{"run", config.LambdaPath}, config.LambdaArgs...)
 		cmd = exec.Command("go", args...)
-
 	default:
 		return nil, fmt.Errorf("unknown run mode: %d", r.Mode)
 	}
 
-	// Set up process group for proper signal handling
+	// needed in order to properly kill lambda process when run with go run
 	cmd.SysProcAttr = &syscall.SysProcAttr{
 		Setpgid: true,
 	}
 
-	// Add the Lambda server port to environment
 	cmd.Env = append(os.Environ(), fmt.Sprintf("_LAMBDA_SERVER_PORT=%s", r.ServerPort))
-
-	// Only show stdout/stderr if debug mode is enabled
-	if r.Debug {
-		cmd.Stdout = os.Stderr
-		cmd.Stderr = os.Stderr
-	} else {
-		cmd.Stdout = nil
-		cmd.Stderr = nil
-	}
 
 	return cmd, nil
 }
 
-// KillProcessGroup kills the entire process group
 func (r *Runner) KillProcessGroup(cmd *exec.Cmd, signal syscall.Signal) error {
 	if cmd.Process == nil {
 		return fmt.Errorf("process not started")
@@ -130,7 +113,6 @@ func (r *Runner) KillProcessGroup(cmd *exec.Cmd, signal syscall.Signal) error {
 		return fmt.Errorf("failed to get process group ID: %w", err)
 	}
 
-	// Kill the entire process group (note the negative sign)
 	if err := syscall.Kill(-pgid, signal); err != nil {
 		return fmt.Errorf("failed to send signal %v: %w", signal, err)
 	}
@@ -138,8 +120,7 @@ func (r *Runner) KillProcessGroup(cmd *exec.Cmd, signal syscall.Signal) error {
 	return nil
 }
 
-// Debugf prints debug messages if debug mode is enabled
-func (r *Runner) Debugf(format string, args ...interface{}) {
+func (r *Runner) Debugf(format string, args ...any) {
 	if r.Debug {
 		fmt.Fprintf(os.Stderr, "[DEBUG] "+format+"\n", args...)
 	}
